+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- MSA visualisation with [pymsaviz](https://github.com/moshi4/pyMSAviz/)
+
 ## [v0.9.2]
 ### Changed
 - Navigation bar links now replaced with a dropdown list of links to sections on the report.